[refactor] discard ajax call for button
