+++ package.json
@@ -9,7 +9,8 @@
         "express-session": "1.17.0",
         "jquery": "^3.4.1",
         "mongodb": "3.5.4",
-        "rainbow-node-sdk": "1.67.1"
+        "rainbow-node-sdk": "1.67.1",
+        "rainbow-web-sdk": "1.67.0"
     },
     "devDependencies": {
         "nodemon": "^2.0.2"
