+++ views/chat.ejs
@@ -13,7 +13,10 @@
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css" integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.5.0/css/all.css" integrity="sha384-B4dIYHKNBt8Bc12p+WXckhzcICo0wtJAoU8YZTY5qE0Id1GSseTk6S+L3BlXeVIU" crossorigin="anonymous">
     <link rel="stylesheet" href="/css/chat.css">
+    <script src="/js/elementsUtils.js"></script>
     <script src="/js/chat.js"></script>
+    <script src="/js/chatBotResponse.js"></script>
+    <script src="/js/chatBotChoice.js"></script>
 </head>
 <body>
     <!--Nav Bar-->
